come back

+++ 1.JavaSyntax/src/com/javarush/task/task10/task1006/Solution.java
@@ -9,11 +9,9 @@ public class Solution {
         double d = (short) 2.50256e2d;
         char c = (short) 'd';
         short s = (short) 2.22;
-        int i = (short) 150000;
-        float f = (short) 0.50f;
+        int i = 150000;
+        float f = 0.50f;
         double result = f + (i / c) - (d * s) - 500e-3;
-
-        //чтобы получился ответ: result: 1000.0.
         System.out.println("result: " + result);
     }
 }
